Do not duplicate case/choice by an explicit `type`
Within the `OMS-elements` list, the model supports multiple "types" or
"classes" of elements -- amplifiers, fiber spans and attenuators, and
possible ROADMs in future. The distinction on "what type of an element"
is currently modeled via two means:

- the `type` leaf which is an `identityref` to a `type-element` identity,
- the `choice` YANG statement and its inner container.

This is redundant, and possibly quite confusing (see the commented-out
`when` stanzas, for example).

YANG's `choice` and `case` statements provide adequate functionality
here because they guarantee that exactly one possibility gets selected.

fixes #21